requirements.txt
Pillow>=8.2.0
piexif>=1.1.3
watchdog>=2.1.2
fastapi>=0.68.0
starlette>=0.18.0
python-magic>=0.4.24
requests>=2.26.0
python-multipart
jinja2>=3.0.0
